+++ .classpath
@@ -24,5 +24,7 @@
 			<attribute name="maven.pomderived" value="true"/>
 		</attributes>
 	</classpathentry>
+	<classpathentry kind="lib" path="C:/Users/sponugoti/Desktop/Selenium Files/selenium-server-standalone-3.141.59.jar"/>
+	<classpathentry kind="lib" path="C:/Users/sponugoti/Desktop/Selenium Files/testng-7.3.0.jar"/>
 	<classpathentry kind="output" path="target/classes"/>
 </classpath>
